refactor/remove login store and make util folder
