+++ src/app/login-feature/login-container/login-container.component.ts
@@ -1,14 +1,5 @@
-import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
-import { LoginComponent } from '../login/login.component';
-import { LoginStore } from '../login.store';
-import { LetModule } from '@ngrx/component';
-import { CommonModule } from '@angular/common';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
-import { ForgotPasswordComponent } from '../forgotPassword/forgotPassword.component';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { ErrorPipe } from '../errors.pipe';
-import { SignUpComponent } from '../signUp/signUp.component';
 import { RouterModule } from '@angular/router';
 
 @Component({
@@ -16,19 +7,7 @@ import { RouterModule } from '@angular/router';
   templateUrl: './login-container.component.html',
   styleUrls: ['login-container.component.scss'],
   standalone: true,
-  providers: [LoginStore],
-  imports: [
-    LoginComponent,
-    LetModule,
-    CommonModule,
-    MatCardModule,
-    ForgotPasswordComponent,
-    SignUpComponent,
-    MatFormFieldModule,
-    MatInputModule,
-    ErrorPipe,
-    RouterModule,
-  ],
+  imports: [MatCardModule, RouterModule],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LoginContainerComponent {}
